+++ docs/core_docs/docs/integrations/callbacks/datadog_tracer.mdx
@@ -4,7 +4,7 @@ sidebar_class_name: node-only
 
 import CodeBlock from "@theme/CodeBlock";
 
-# Datadog LLM Observability
+# Datadog LLM Observability Tracer Callback
 
 :::warning
 LLM Observability is in public beta, and its API is subject to change.
@@ -30,6 +30,6 @@ npm install @langchain/community
 
 ## Usage
 
-import UsageExample from "@examples/tracers/datadog.ts";
+import UsageExample from "@examples/callbacks/datadog.ts";
 
 <CodeBlock language="typescript">{UsageExample}</CodeBlock>
